refactor to node_info_loader
